Move animation classes to Mapsui.Animations
