+++ Mapsui/Animations/AnimationEntry.cs
@@ -1,7 +1,6 @@
-﻿using Mapsui.Animations;
-using System;
+﻿using System;
 
-namespace Mapsui.Utilities;
+namespace Mapsui.Animations;
 
 public class AnimationEntry<T>
 {
