+++ Mapsui/UI/MouseWheelAnimation.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Mapsui.Animations;
 using Mapsui.Utilities;
 
 namespace Mapsui.UI;
