+++ Mapsui/ViewportAnimations/FlingAnimation.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using Mapsui.Animations;
 using Mapsui.Extensions;
-using Mapsui.Utilities;
 
 namespace Mapsui.ViewportAnimations;
