+++ Mapsui/ViewportAnimations/ViewportStateAnimation.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using Mapsui.Animations;
-using Mapsui.Utilities;
 
 namespace Mapsui.ViewportAnimations;
