+++ Mapsui/ViewportAnimations/ZoomAroundLocationAnimation.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Mapsui.Animations;
 using Mapsui.Utilities;
 
 namespace Mapsui.ViewportAnimations;
