+++ Samples/Mapsui.Samples.Common/Maps/Animations/ViewportCenterAndZoomAnimationSample.cs
@@ -1,7 +1,6 @@
-﻿using Mapsui.Extensions;
+﻿using Mapsui.Animations;
+using Mapsui.Extensions;
 using Mapsui.Tiling;
-using Mapsui.UI;
-using Mapsui.Utilities;
 using Mapsui.Widgets;
 using Mapsui.Widgets.ScaleBar;
 using Mapsui.Widgets.Zoom;
